refactor std error and some formatting
